chore: fix copyrights and refactor test package
