+++ contract/src/main/kotlin/com/ritense/valtimo/contract/http/ValtimoHttpRestTemplateConfigurationProperties.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2023 Ritense BV, the Netherlands.
+ * Copyright 2015-2024 Ritense BV, the Netherlands.
  *
  * Licensed under EUPL, Version 1.2 (the "License");
  * you may not use this file except in compliance with the License.
